fix: resolve ambiguous SetAttribute overload for CrimeWatch timers
Cast all timer values to int64 explicitly to fix build error on
compilers where 'long long' is ambiguous between int/uint32/int64.

+++ src/eve-server/system/CrimeWatch.cpp
@@ -65,13 +65,13 @@ void CrimeWatch::Process()
     if (m_aggressionTimer.Enabled() and m_aggressionTimer.Check(false)) {
         m_aggressionTimer.Disable();
         if (m_client->GetChar())
-            m_client->GetChar()->SetAttribute(ATTR_AGGRESSION_TIMER, 0LL, true);
+            m_client->GetChar()->SetAttribute(ATTR_AGGRESSION_TIMER, int64(0), true);
     }
     if (m_criminalTimer.Enabled()) m_criminalTimer.Check();
     if (m_weaponTimer.Enabled() and m_weaponTimer.Check(false)) {
         m_weaponTimer.Disable();
         if (m_client->GetChar())
-            m_client->GetChar()->SetAttribute(ATTR_WEAPON_TIMER, 0LL, true);
+            m_client->GetChar()->SetAttribute(ATTR_WEAPON_TIMER, int64(0), true);
     }
     if (m_limitedEngagementTimer.Enabled()) m_limitedEngagementTimer.Check();
 }
@@ -95,7 +95,7 @@ void CrimeWatch::OnWeaponFired()
     m_weaponTimer.Start(60000);
     if (m_client->GetChar()) {
         int64 endTime = static_cast<int64>(GetFileTimeNow()) + 60000LL * 1000LL;
-        m_client->GetChar()->SetAttribute(ATTR_WEAPON_TIMER, endTime, true);
+        m_client->GetChar()->SetAttribute(ATTR_WEAPON_TIMER, int64(endTime), true);
     }
 }
 
@@ -107,8 +107,8 @@ void CrimeWatch::OnLooting()
     m_aggressionTimer.Start(sConfig.crime.AggFlagTime * 1000);
     if (m_client->GetChar()) {
         int64 now = static_cast<int64>(GetFileTimeNow());
-        m_client->GetChar()->SetAttribute(ATTR_WEAPON_TIMER, now + 60000LL * 1000LL, true);
-        m_client->GetChar()->SetAttribute(ATTR_AGGRESSION_TIMER, now + sConfig.crime.AggFlagTime * 1000LL * 1000LL, true);
+        m_client->GetChar()->SetAttribute(ATTR_WEAPON_TIMER, int64(now + 60000LL * 1000LL), true);
+        m_client->GetChar()->SetAttribute(ATTR_AGGRESSION_TIMER, int64(now + sConfig.crime.AggFlagTime * 1000LL * 1000LL), true);
     }
     // -0.2 security penalty (already applied in InventoryBound::Add)
 }
@@ -155,7 +155,7 @@ void CrimeWatch::OnAggression(Client* pTarget, float systemSecRating)
     m_aggressionTimer.Start(sConfig.crime.AggFlagTime * 1000);
     if (m_client->GetChar()) {
         int64 endTime = static_cast<int64>(GetFileTimeNow()) + sConfig.crime.AggFlagTime * 1000LL * 1000LL;
-        m_client->GetChar()->SetAttribute(ATTR_AGGRESSION_TIMER, endTime, true);
+        m_client->GetChar()->SetAttribute(ATTR_AGGRESSION_TIMER, int64(endTime), true);
     }
 
     // Highsec: criminal act + CONCORD response + kill right grant
